requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.1
cryptography==44.0.2
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
numpy==1.26.4
pandas==2.2.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
langchain-core>=0.3.15,<0.4.0
langchain>=0.3.7
langchain-openai==0.2.6
mcp>=1.26.0
scikit-learn==1.5.2
scipy==1.13.1
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
